+++ vite.config.umd.ts
@@ -20,6 +20,7 @@ export default defineConfig({
       fileName: () => "index.js",
     },
     minify: "terser",
+    terserOptions: { maxWorkers: 1 },
     rollupOptions: {
       external: ["react", "dayjs"],
       output: {